requirements.txt
Flask
python-dotenv
openai
whisper
ffmpeg-python